Initial Commit
Added two Java files that contain the source code for solving the Viagogo Developer Test. Events.java contains the details required to create an Object called Event that stores an ID, X and Y axis, number of tickets and a ticket price. ClosestUserEvents.java contains the code required to generate a series of random events on a X and Y axis as well prompt the user for their own coordinates. It then used the Manhattan distance algorithm to calculate the distance from the user to the event. It then finds the five closest events to the user and lists their details.